Move component wrapper to avoid circular dependencies
