+++ projects/storefrontlib/src/cms-structure/page/component/page-component.module.ts
@@ -1,0 +1,11 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { ComponentWrapperDirective } from './component-wrapper.directive';
+
+@NgModule({
+  imports: [CommonModule],
+  providers: [],
+  declarations: [ComponentWrapperDirective],
+  exports: [ComponentWrapperDirective]
+})
+export class PageComponentModule {}
