+++ projects/storefrontlib/src/cms-structure/page/index.ts
@@ -1,2 +1,7 @@
+// slot
 export * from './slot/page-slot.module';
 export * from './slot/page-slot.component';
+
+// component
+export * from './component/page-component.module';
+export * from './component/component-wrapper.directive';
