+++ projects/storefrontlib/src/cms-structure/page/slot/page-slot.module.ts
@@ -2,10 +2,10 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { PageSlotComponent } from './page-slot.component';
 import { OutletModule } from 'projects/storefrontlib/src/lib/outlet';
-import { CmsModule } from '../../../lib/cms/index';
+import { PageComponentModule } from '../component/page-component.module';
 
 @NgModule({
-  imports: [CommonModule, OutletModule, CmsModule],
+  imports: [CommonModule, OutletModule, PageComponentModule],
   providers: [],
   declarations: [PageSlotComponent],
   exports: [PageSlotComponent]
