--- projects/storefrontlib/src/lib/cms/components/index.ts
@@ -1,5 +1,0 @@
-export * from './component-wrapper.directive';
-
-import { ComponentWrapperDirective } from './component-wrapper.directive';
-
-export const components: any[] = [ComponentWrapperDirective];
